delete & update HQL added

+++ src/main/java/com/excilys/cdb/dao/CompanyDAOInterface.java
@@ -23,6 +23,6 @@ public interface CompanyDAOInterface<T> {
 	 * @throws IOException 
 	 * @throws DataBaseException 
 	 */
-	public abstract void delete(int id) throws IOException, DataBaseException;
+	public abstract void delete(long id) throws IOException, DataBaseException;
 	
 }
